+++ src/components/Menu/index.ts
@@ -1,0 +1,14 @@
+export * from "./Menu.types";
+export {
+  MenuRoot,
+  MenuTrigger,
+  MenuPortal,
+  MenuBackdrop,
+  MenuPositioner,
+  MenuPopup,
+  MenuArrow,
+  MenuItem,
+  MenuSeparator,
+  MenuGroup,
+  MenuGroupLabel,
+} from "./Menu";