+++ src/components/index.ts
@@ -56,3 +56,27 @@ export * from "../hooks/useInputBase";
 // System
 export * from "../system/styleEngine";
 export * from "../hooks/useStyleResolver";
+
+export {
+  // Menu,
+  MenuRoot,
+  MenuTrigger,
+  // MenuPortal,
+  MenuBackdrop,
+  MenuPositioner,
+  MenuPopup,
+  MenuArrow,
+  MenuItem,
+  MenuSeparator,
+  MenuGroup,
+  MenuGroupLabel,
+  // MenuRadioGroup,
+  // MenuRadioItem,
+  // MenuRadioItemIndicator,
+  // MenuCheckboxItem,
+  // MenuCheckboxItemIndicator,
+  // MenuSubmenuRoot,
+  // MenuSubmenuTrigger,
+} from "./Menu";
+
+export * from "./Menu/Menu.types";